+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.2.1 | TBD
 
 - Added new option: `/playing` can now receive an updating behavior parameter
+- Added album name to `/playing` embed
 - Fixed a bug where uncached guilds would panic the bot
 - Fixed small issue with embed styling
 - Updated to Rust 1.80.1 (from 1.79.0)